refactor (windowDesc): renamed file with class windowDesk to window_desc, consequently changing includes and CMakeLists.txt
