+++ source/desk_up_window_backend/desk_up_window_device.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <string>
-#include "windowDesc.h"
+#include "window_desc.h"
 
 struct DeskUpWindowDevice{
     //here goes all the generic functions that all the backends have
