+++ source/desk_up_window_backend/window_desc/CMakeLists.txt
@@ -2,4 +2,4 @@ add_library(window_desc_library STATIC)
 
 target_include_directories(window_desc_library PUBLIC ${CMAKE_SOURCE_DIR}/source/desk_up_window_backend)
 
-target_sources(window_desc_library PUBLIC ${CMAKE_CURRENT_LIST_DIR}/windowDesc.cc ${CMAKE_CURRENT_LIST_DIR}/windowDesc.h)
+target_sources(window_desc_library PUBLIC ${CMAKE_CURRENT_LIST_DIR}/window_desc.cc ${CMAKE_CURRENT_LIST_DIR}/window_desc.h)
